Rework `Gemfile.local` --> `.pdqtest/Gemfile`
Turns out there is just no way to win with fighting PDK over ownership
of `Gemfile`/`Gemfile.lock`. My attempts to make this work just about
worked on a workstation with a lot of effort but I was not able to make
this work at all in CI. The reson for this is that PDK changes the
bundle contents while Ruby is running and it does *not* like this one
bit.

Turns out there was a real simple soultion I was overlooking - just have
two `Gemfile`s - this patch introduces `.pdqtest/Gemfile` which is where
we store all our gem related voodoo from now on. This allows complete
separation from PDK and no more trying to "each each other's dinner off
the plate"