Inplace execution
Some CI systems, its just too painful to run docker-in-docker, so
we now allow _inplace_ execution inside the system running PDQTest.

Don't run this on a computer system you care about! Infact, don't
run it at all. The intention of this feature is for it to work
transparently through the various CI configuration files.

Currently targeted: Bitbucket Pipelines and Appveyor